Refactoring to separate out utility functions and fixing some missing cases
